requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Pillow==10.0.1
gunicorn==21.2.0
whitenoise==6.5.0
django-filter==23.3
drf-yasg==1.21.7
grpcio==1.59.0
grpcio-tools==1.59.0
django-allauth==0.54.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
celery==5.3.4
redis==5.0.1